+++ gui/datumaro_gui/components/multiple/tabs/transform.py
@@ -15,6 +15,7 @@ from datumaro_gui.utils.dataset.data_loader import MultipleDatasetHelper
 from datumaro_gui.utils.dataset.info import get_category_info, get_subset_info
 from datumaro_gui.utils.drawing import Dashboard, Pie, Radar
 from datumaro_gui.utils.drawing.css import box_style
+from datumaro_gui.utils.page import check_image_backend
 from streamlit import session_state as state
 from streamlit_elements import elements
 
@@ -697,6 +698,8 @@ def render_dataset_management_section(
 
 
 def main():
+    check_image_backend(state.get("IMAGE_BACKEND"))
+
     data_helper_1: MultipleDatasetHelper = state["data_helper_1"]
     data_helper_2: MultipleDatasetHelper = state["data_helper_2"]
     uploaded_file_1 = state["uploaded_file_1"]