+++ gui/datumaro_gui/components/single/tabs/validate.py
@@ -7,11 +7,13 @@ from types import SimpleNamespace
 import streamlit as st
 from datumaro_gui.utils.dataset.data_loader import SingleDatasetHelper
 from datumaro_gui.utils.drawing import Bar, Dashboard, DataGrid
+from datumaro_gui.utils.page import check_image_backend
 from streamlit import session_state as state
 from streamlit_elements import elements
 
 
 def main():
+    check_image_backend(state.get("IMAGE_BACKEND"))
     data_helper: SingleDatasetHelper = state["data_helper"]
     tasks = ["classification", "detection", "segmentation"]
     if "task" not in state: