+++ src/datumaro/components/dataset.py
@@ -441,7 +441,7 @@ class Dataset(IDataset):
         Returns: self
 
         Example:
-            - (`filter_annotations=True`) This is an example of filtering
+            - (`filter_annotations=False`) This is an example of filtering
                 dataset items with images larger than 1024 pixels::
 
                 from datumaro.components.media import Image
@@ -450,8 +450,7 @@ class Dataset(IDataset):
                     h, w = item.media_as(Image).size
                     return h > 1024 or w > 1024
 
-                filtered = UserFunctionDatasetFilter(
-                    extractor=dataset, filter_func=filter_func)
+                filtered = dataset.filter(filter_func=filter_func, filter_annotations=False)
                 # No items with an image height or width greater than 1024
                 filtered_items = [item for item in filtered]
 
@@ -473,8 +472,7 @@ class Dataset(IDataset):
                     # Accept Bboxes smaller than 50% of the image size
                     return bbox_size < 0.5 * image_size
 
-                filtered = UserFunctionAnnotationsFilter(
-                    extractor=dataset, filter_func=filter_func)
+                filtered = dataset.filter(filter_func=filter_func, filter_annotations=True)
                 # No bounding boxes with a size greater than 50% of their image
                 filtered_items = [item for item in filtered]
         """