+++ tests/unit/gui/single/test_explore.py
@@ -5,7 +5,6 @@
 import os
 from unittest import TestCase
 
-import pytest
 from streamlit.testing.v1 import AppTest
 
 from datumaro.components.visualizer import Visualizer
@@ -27,9 +26,11 @@ def run_explore():
     from gui.datumaro_gui.components.single import tabs
     from gui.datumaro_gui.utils.dataset.data_loader import SingleDatasetHelper
     from gui.datumaro_gui.utils.dataset.state import reset_state, single_state_keys
+    from gui.datumaro_gui.utils.page import init_func
 
     from tests.utils.assets import get_test_asset_path
 
+    init_func(state.get("IMAGE_BACKEND", None))
     reset_state(single_state_keys, state)
 
     dataset_dir = get_test_asset_path("explore_dataset")
@@ -180,7 +181,6 @@ class ExploreTest(TestCase):
         # Check updated query status
         assert len(at.session_state.explore_queries) == 1
 
-    @pytest.mark.xfail(reason="Cannot copy contextvar to thread")
     @mark_requirement(Requirements.DATUM_GENERAL_REQ)
     def test_explore_check_dataset_image(self):
         """Test if label method of explore are worked out correctly."""
@@ -265,7 +265,6 @@ class ExploreTest(TestCase):
         # TODO
         """Test if user image method of explore are worked out correctly."""
 
-    @pytest.mark.xfail(reason="Cannot copy contextvar to thread")
     @mark_requirement(Requirements.DATUM_GENERAL_REQ)
     def test_explore_check_text(self):
         """Test if text method of explore are worked out correctly."""